+++ tests/logging_callbacks.hpp
@@ -1,6 +1,7 @@
 #pragma once
 
 #include <periple/core/moves/append_move.hpp>
+#include <periple/core/moves/dp_move.hpp>
 
 #include <string>
 #include <vector>
@@ -20,6 +21,17 @@ struct LoggingCallbacks {
 	template <typename CityT>
 	void move_prepare(const AppendMove<CityT>&) const {
 		log.push_back("move_prepare");
+	}
+
+	template <typename CityT>
+	bool move_filter(const DPMove<CityT>&) const {
+		log.push_back("dp_move_filter");
+		return true;
+	}
+
+	template <typename CityT>
+	void move_prepare(const DPMove<CityT>&) const {
+		log.push_back("dp_move_prepare");
 	}
 };